Breaking change: replaced IsSucceeded extension method for classic searches with an explicit IsSucceeded property.
Deals with a minor bug - if the target node is a default struct it would have returned a false negative. A bit (okay, alot) of an edge case, but possible. Consider for example a graph in which nodes are just an int-valued index into some underlying static data store, and a search for which the target node happens to have the index zero.

(also some test fixes)